fix: Change the format with props and change css properties about border
