+++ app/vote-edit/page.tsx
@@ -1,8 +1,64 @@
 import React from "react";
 import EditOptionArea from "./_components/EditOptionArea";
+import { ColumnsType } from "../edit-course/_components/DragAndDropArea";
+
+const initialColumns: ColumnsType = {
+  course: {
+    id: "course",
+    list: {
+      food: [{ globalIndex: 0, title: "음식점", type: "food", icon: "🍔" }],
+      dessert: [{ globalIndex: 1, title: "카페", type: "dessert", icon: "🥨" }],
+      beer: [
+        { globalIndex: 2, title: "술 1차", type: "dessert", icon: "🥨" },
+        { globalIndex: 3, title: "술 2차", type: "dessert", icon: "🥨" },
+      ],
+      play: [{ globalIndex: 4, title: "놀거리", type: "play", icon: "🥨" }],
+    },
+  },
+};
+
+const placesInfo = [
+  {
+    place: "옥소반 상암점",
+    link: "abcd",
+    rating: "4.01",
+    reviewCount: 433,
+    images: ["/png/food.png"],
+  },
+  {
+    place: "스타벅스 강남점",
+    link: "efgh",
+    rating: "4.5",
+    reviewCount: 1200,
+    images: ["/png/food.png"],
+  },
+  {
+    place: "맥도날드 홍대점",
+    link: "ijkl",
+    rating: "3.8",
+    reviewCount: 530,
+    images: ["/png/food.png"],
+  },
+  {
+    place: "빕스 여의도점",
+    link: "mnop",
+    rating: "4.2",
+    reviewCount: 870,
+    images: ["/png/food.png"],
+  },
+  {
+    place: "이디야 커피 신촌점",
+    link: "qrst",
+    rating: "4.0",
+    reviewCount: 300,
+    images: ["/png/food.png"],
+  },
+];
 
 const VoteEditPage = () => {
-  return <EditOptionArea />;
+  return (
+    <EditOptionArea initialColumns={initialColumns} placesInfo={placesInfo} />
+  );
 };
 
 export default VoteEditPage;
